* commit '8b98e30e22824c4bb6384c1357f76903aebf12d5': (1210 commits)
  还原
  蠢到爆的大小写。。。。
  Revert "Revert "组合Table的Relation""
  Revert "组合Table的Relation"
  组合Table的Relation
  bug
  计算公式值为空的时候处理
  update
  文件
  bug
  业务包分组删除按钮样式
  业务包分组管理提示修改（交互）
  数据连接schema，修改excel数据集
  bug
  业务包图标样式（输入边框）
  业务包图标样式修改（交互）
  范围面积重定义 bug
  bug fix
  数据连接
  汇总表排序bug
  ...

+++ fbi_engine_base/src/com/fr/bi/cal/stable/cube/file/AbstractSingleColumnFile.java
@@ -15,7 +15,7 @@ import com.fr.bi.stable.io.newio.NIOReader;
 import com.fr.bi.stable.io.newio.NIOWriter;
 import com.fr.bi.stable.io.newio.SingleUserNIOReadManager;
 import com.fr.bi.stable.io.sortlist.ISortNIOReadList;
-import com.fr.bi.stable.relation.BITableSourceRelation;
+import com.finebi.cube.relation.BITableSourceRelation;
 import com.fr.bi.stable.structure.array.ArrayKey;
 import com.finebi.cube.api.ICubeColumnIndexReader;
 import com.fr.bi.stable.utils.file.BIFileUtils;
@@ -81,7 +81,7 @@ public abstract class AbstractSingleColumnFile<T> extends LinkIndexFile implemen
     }
 
     @Override
-    public void addDataValue(long row, T value) {
+    public void addDataValue(int row, T value) {
         detail.createNIOWriter().add(row, value);
     }
